Dataabse conncetion and query  function updated Configuration file created for database connection
