+++ config.php
@@ -1,0 +1,11 @@
+<?php
+return [
+    'database'=>[
+        'host'  => 'localhost',
+        'port'  => 3306,
+        'dbname'=>'myapp',
+        'charset' => 'utf8mb4'
+    ],
+    
+    // services or api to register
+];
